creating component to add details
