+++ src/app/app-routing.module.ts
@@ -9,6 +9,8 @@ import { AuthGuardService } from './shared/service/authguard.service';
 import { SigninComponent } from './signin/signin.component';
 import { InvestmentEditComponent } from './home/investments/investment-edit/investment-edit.component';
 import { TransactionsComponent } from './home/transactions/transactions.component';
+import { TransactionsAddComponent } from './home/transactions/transactions-add/transactions-add.component';
+import { TransactionsAddDetailsComponent } from './home/transactions/transactions-add-details/transactions-add-details.component';
 
 const routes: Routes = [
   { path: 'signin', component: SigninComponent },
@@ -20,7 +22,12 @@ const routes: Routes = [
       { path: 'investments/:id/edit', component: InvestmentEditComponent },
       { path: 'investments', component: InvestmentsComponent },
       { path: 'members', component: MembersComponent },
-      { path: 'transactions', component: TransactionsComponent }
+      {
+        path: 'transactions', component: TransactionsComponent, children: [
+          { path: '', component: TransactionsAddComponent, pathMatch: 'full' },
+          { path: 'add/:id/details', component: TransactionsAddDetailsComponent }
+        ]
+      }
     ]
   }
 ];
