+++ src/app/home/transactions/transactions-add-details/transactions-add-details.component.html
@@ -1,0 +1,1 @@
+<p>transactions-add-details works!</p>
