+++ src/app/home/transactions/transactions-add-details/transactions-add-details.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-transactions-add-details',
+  templateUrl: './transactions-add-details.component.html'
+})
+export class TransactionsAddDetailsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
