+++ src/app/home/transactions/transactions.component.html
@@ -1,22 +1,1 @@
-<div class="row">
-    <div class="col-xs-12">
-        <form [formGroup]="transactionsForm" (ngSubmit)="onSubmit(fD)" #fD="ngForm">
-            <div class="row">
-                <div class="col-xs-12">
-                    <mat-form-field class="w-100" appearance="outline">
-                        <mat-label>Paste Transactions</mat-label>
-                        <textarea type="text" matInput formControlName="transactions" cdkTextareaAutosize></textarea>
-                        <mat-error>Please enter a valid details</mat-error>
-                    </mat-form-field>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-xs-12">
-                    <button type="submit" mat-raised-button color="primary"
-                        [disabled]="!this.transactionsForm.valid">Add Transactions</button>
-                </div>
-            </div>
-        </form>
-    </div>
-</div>
+<router-outlet></router-outlet>
